personal-calendar/requirements.txt
fastapi
uvicorn
psycopg2-binary
jinja2 